backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.30.1
httpx<0.28
python-dotenv==1.0.1
pydantic==2.6.4
python-multipart==0.0.9
pytest==8.2.2
